Renamed HashFunctionTests to match class under test ChecksumAlgorithm
